Renamed moc_loops to moc_globe_atl
