+++ src/geocat/f2py/__init__.py
@@ -3,6 +3,6 @@ from .dpres_plevel_wrapper import (dpres_plevel)
 from .errors import *
 from .linint2_wrapper import (linint1, linint2, linint2pts)
 from .missing_values import (py2fort_msg, fort2py_msg)
-from .moc_loops_wrapper import (moc_globe_atl)
+from .moc_globe_atl_wrapper import (moc_globe_atl)
 from .rcm2rgrid_wrapper import (rcm2rgrid, rgrid2rcm)
 from .triple2grid_wrapper import (grid_to_triple, triple_to_grid)
